librbd_sys::rbd_create3

pub unsafe extern fn rbd_create3(io: rados_ioctx_t, name: *const c_char, size: uint64_t, features: uint64_t, order: *mut c_int, stripe_unit: uint64_t, stripe_count: uint64_t) -> c_int

create new rbd image

The stripe_unit must be a factor of the object size (1 << order). The stripe_count can be one (no intra-object striping) or greater than one. The RBD_FEATURE_STRIPINGV2 must be specified if the stripe_unit != the object size and the stripe_count is != 1.

Arguments

io ioctx

name image name

size image size in bytes

features initial feature bits

order object/block size, as a power of two (object size == 1 << order)

stripe_unit stripe unit size, in bytes.

stripe_count number of objects to stripe over before looping @return 0 on success, or negative error code